qianfan-ocr-test/requirements.txt
torch==2.11.0
torchvision==0.26.0
transformers==5.5.4
pillow==12.2.0
accelerate==1.13.0
einops==0.8.2
timm==1.0.26
sentencepiece==0.2.1
